chore(LotteryTest): 잘못된 클래스에 있던 테스트 코드를 옮긴다
