+++ src/test/kotlin/model/lottery/LotteriesTest.kt
@@ -1,9 +1,6 @@
 package model.lottery
 
-import model.Quantity
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.Assertions.assertFalse
-import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
 
 class LotteriesTest {
@@ -45,28 +42,5 @@ class LotteriesTest {
                 ),
             ),
         )
-    }
-
-    @Test
-    fun `로또 번호를 당첨 번호와 비교하여 서로 같은 번호를 센다`() {
-        val lottery = Lottery.of(3, 6, 9, 12, 15, 18)
-        val winningLottery = Lottery.of(2, 4, 6, 8, 10, 12)
-
-        val actual = lottery.compareLottery(winningLottery)
-        assertThat(actual).isEqualTo(Quantity(2))
-    }
-
-    @Test
-    fun `로또 번호가 보너스 번호를 포함한다면 참이다`() {
-        val lottery = Lottery.of(4, 8, 12, 16, 20, 24)
-        val bonusNumber = LotteryNumber(20)
-        assertTrue(lottery.contains(bonusNumber))
-    }
-
-    @Test
-    fun `로또 번호가 보너스 번호를 포함하지 않는다면 거짓이다`() {
-        val lottery = Lottery.of(2, 4, 6, 8, 10, 12)
-        val bonusNumber = LotteryNumber(21)
-        assertFalse(lottery.contains(bonusNumber))
     }
 }
